requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.
pbr!=2.1.0,>=2.0.0 # Apache-2.0

beautifulsoup4>=4.6.0 # MIT
Django<2;python_version<'3.0'  # BSD
Django<2.1;python_version>='3.0'  # BSD
django-formtools>=1.0 # BSD
iso8601>=0.1.11 # MIT
six>=1.10.0 # MIT
python-muranoclient>=0.8.2 # Apache-2.0
pytz>=2013.6 # MIT
PyYAML>=3.12 # MIT
yaql>=1.1.3 # Apache 2.0 License
castellan>=0.16.0 # Apache-2.0

oslo.log>=3.36.0 # Apache-2.0
semantic-version>=2.3.1 # BSD

horizon>=13.0.0 # Apache-2.0

# message extraction
Babel!=2.4.0,>=2.3.4 # BSD
django-babel>=0.6.2 # BSD
